Battleship+ project
This project involves the coding of a simple Python program that allows a User to play an extended version of Battleship with the Computer. The User is asked to place his/her Ships on the board according 3D coordinates (row, column, depth) and orientation (vertical or horizontal), followed by the Computer.

The User is then asked a set of 3D coordinates, which represents an attack with an impact that is 3 x 3 in area, on the Computer’s Ships, and the program will report if it is a hit, a miss, and if any Ships are sunk. After the User’s turn, the Computer will then randomly select a set of coordinates and the Computer reports the results. The game will end when either the User or Computer’s Ships are all sunk.